Update README.md
## Introduction:
The Covid_19 pandemic has wreaked havoc and led to the dramatic loss of life and
livelihood. Its impact continues to affect the way in which we live and interact. In
this project,I analyse data related to COVID-19 cases as recorded during
January 2020 when COVID-19 was first reported. The data for this project is provided as a CSV file.